+++ src/main/kotlin/racingcar/domain/racing/NumberPicker.kt
@@ -2,13 +2,11 @@ package racingcar.domain.racing
 
 import camp.nextstep.edu.missionutils.Randoms
 
-class NumberPicker {
+object NumberPicker {
+    private const val MIN: Int = 0
+    private const val MAX: Int = 9
+
     fun getRandomNumber(): Int {
         return Randoms.pickNumberInRange(MIN, MAX)
-    }
-
-    companion object {
-        private const val MIN: Int = 0
-        private const val MAX: Int = 9
     }
 }
