+++ src/test/kotlin/racingcar/domain/car/NameParserTest.kt
@@ -1,13 +1,12 @@
-package racingcar.domain.util
+package racingcar.domain.car
 
-import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.DisplayNameGeneration
 import org.junit.jupiter.api.DisplayNameGenerator
 import org.junit.jupiter.api.Test
-import racingcar.util.Parser
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores::class)
-class ParserTest {
+class NameParserTest {
 
     @Test
     fun 입력값을_쉼표를_기준으로_올바르게_분리한다() {
@@ -15,11 +14,11 @@ class ParserTest {
         val input = "pobi,woni , jun"
 
         // when
-        val names = Parser.splitByComma(input)
+        val names = NameParser.splitByComma(input)
 
         // then
-        assertEquals(names[0], "pobi")
-        assertEquals(names[1], "woni")
-        assertEquals(names[2], "jun")
+        Assertions.assertEquals(names[0], "pobi")
+        Assertions.assertEquals(names[1], "woni")
+        Assertions.assertEquals(names[2], "jun")
     }
 }
